Refactor bookmarks_view
Add "contract" to bookmarks_view
