+++ app/src/main/java/org/kiwix/kiwixmobile/bookmarks_view/BookmarksActivity.java
@@ -46,6 +46,7 @@ import android.widget.ListView;
 import org.kiwix.kiwixmobile.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.base.BaseActivity;
+import org.kiwix.kiwixmobile.bookmarks_view.contract.BookmarksViewCallback;
 import org.kiwix.kiwixmobile.settings.KiwixSettingsActivity;
 import org.kiwix.kiwixmobile.utils.SharedPreferenceUtil;
